Add combak comment about blue constraint solver.
It's tempting to want to have a global equality solver at this point.
The tradeoff is debuggability/locality, which I believe to be an
important property. In the current version, a Bluefish spec is pretty
locally specified. That means it's easy to modify and run. It's easy to
put parts in or take them out because there are a lot of little
boundaries between layouts (each component is a layout boundary). If we
have a global equality solver, then we open the door to potentially
weird non-local behavior. And we lose a predictable execution order.